# openapi.json
{"openapi":"3.0.1","info":{"title":"open-api doc","version":"v1"},"paths":{"/":{"get":{"operationId":"index","tags":["IndexController"],"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"type":"string"}}}}},"parameters":[]}},"/v3/api-docs":{"get":{"operationId":"apiDocs","tags":["IndexController"],"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"type":"string"}}}}},"parameters":[]}},"/healthcheck":{"get":{"operationId":"healthcheck","tags":["IndexController"],"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponseOfVoid"}}}}},"parameters":[]}},"/orders":{"post":{"operationId":"saveOrder","tags":["OrderController"],"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}}}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderApiRequest"}}}},"parameters":[]}},"/orders/{id}":{"get":{"operationId":"getOrder","tags":["OrderController"],"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}}}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}]},"put":{"operationId":"updateOrder","tags":["OrderController"],"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}}}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderApiRequest"}}}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}]}}},"components":{"schemas":{"ApiResponseOfVoid":{"type":"object","properties":{"successful":{"type":"boolean"},"message":{"type":"string"},"code":{"type":"string"}}},"Order":{"type":"object","properties":{"id":{"type":"number"},"createdAt":{"type":"string","format":"date-time"},"amount":{"type":"number"},"customerId":{"type":"string"},"createdBy":{"type":"string"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItem"},"uniqueItems":false}}},"OrderItem":{"type":"object","properties":{"id":{"type":"number"},"productId":{"type":"string"},"amount":{"type":"number"}}},"OrderApiRequest":{"type":"object","properties":{"customerId":{"type":"string"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemApiRequest"},"uniqueItems":false}}},"OrderItemApiRequest":{"type":"object","properties":{"productId":{"type":"string"},"amount":{"type":"number"}}}}}}